move registration business logic to register.js
