+++ librefit-web/src/lib/components/modal/LoginModal.svelte
@@ -1,0 +1,23 @@
+<script>
+    import {getModalStore} from '@skeletonlabs/skeleton';
+
+    let value;
+
+    const modalStore = getModalStore();
+
+    const onSubmit = () => {
+        if ($modalStore[0].response) {
+            $modalStore[0].response({
+                detail: {
+                    value: value
+                }
+            });
+        }
+    }
+
+    const onCancel = () => {
+        if ($modalStore[0].response) {
+            $modalStore[0].response(undefined);
+        }
+    }
+</script>
